chore(DataMapper): reorganize Document/actions/ file structure
Fixes: https://github.com/KaotoIO/kaoto/issues/3210

- Move field context menu components into FieldContextMenu sub directory
