+++ packages/ui/src/components/Document/SourceDocumentNode.tsx
@@ -7,8 +7,8 @@ import { TreeUIService } from '../../services/visualization/tree-ui.service';
 import { VisualizationService } from '../../services/visualization/visualization.service';
 import { VisualizationUtilService } from '../../services/visualization/visualization-util.service';
 import { useDocumentTreeStore } from '../../store';
+import { withFieldContextMenu } from './actions/FieldContextMenu';
 import { OverrideIndicator } from './actions/FieldOverride/OverrideIndicator';
-import { withFieldContextMenu } from './actions/withFieldContextMenu';
 import { handleNodeKeyDown } from './document-node.utils';
 import { NodeContainer } from './NodeContainer';
 import { BaseNode } from './Nodes/BaseNode';
