+++ packages/ui/src/components/Document/actions/FieldContextMenu/index.ts
@@ -1,0 +1,6 @@
+export type { IFieldContextMenuProps } from './FieldContextMenu';
+export { FieldContextMenu } from './FieldContextMenu';
+export { useContextMenuState } from './useContextMenuState';
+export { withFieldContextMenu } from './withFieldContextMenu';
+export type { WrapperSelectionModalProps } from './WrapperSelectionModal';
+export { WrapperSelectionModal } from './WrapperSelectionModal';
